bugfix, test improvements
