+++ test/visitors/declare_opaque_type.test.ts
@@ -9,5 +9,10 @@ pluginTester({
       code: `declare opaque type TimeoutID;`,
       output: `declare type TimeoutID = unknown;`,
     },
+    {
+      title: 'declare opaque type',
+      code: `declare opaque type TimeoutID: number;`,
+      output: `declare type TimeoutID = unknown;`,
+    },
   ],
 });
